@@ -123,60 +123,61 @@
     "@babel/plugin-syntax-import-meta": "7.10.4",
     "@babel/preset-env": "7.20.2",
     "@babel/preset-react": "7.18.6",
-    "@babel/preset-typescript": "7.18.6",
+    "@babel/preset-typescript": "7.21.0",
     "@electron/notarize": "1.2.3",
     "@lavamoat/allow-scripts": "2.3.0",
     "@lavamoat/preinstall-always-fail": "1.0.0",
-    "@ledgerhq/hw-transport-mocker": "6.27.10",
+    "@ledgerhq/hw-transport-mocker": "6.27.12",
     "@parcel/config-default": "2.8.3",
     "@parcel/core": "2.8.3",
     "@parcel/transformer-css": "2.8.3",
     "@parcel/transformer-stylus": "2.8.3",
     "@parcel/transformer-typescript-tsc": "2.8.3",
-    "@sentry/types": "7.33.0",
-    "@testing-library/dom": "8.20.0",
-    "@testing-library/react": "13.4.0",
+    "@sentry/types": "7.42.0",
+    "@testing-library/dom": "9.0.1",
+    "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.4.3",
     "@types/cookie": "0.5.1",
+    "@types/deep-equal": "^1.0.1",
     "@types/get-pixels": "3.3.2",
     "@types/hdkey": "2.0.1",
-    "@types/jest": "29.2.6",
-    "@types/node": "18.11.18",
+    "@types/jest": "29.4.0",
+    "@types/node": "18.14.6",
     "@types/node-fetch": "2.6.2",
     "@types/node-hid": "1.3.1",
     "@types/tar-fs": "2.0.1",
-    "@types/uuid": "9.0.0",
+    "@types/uuid": "9.0.1",
     "@types/ws": "8.5.4",
-    "@typescript-eslint/eslint-plugin": "5.48.1",
-    "@typescript-eslint/parser": "5.48.1",
-    "electron": "22.0.3",
+    "@typescript-eslint/eslint-plugin": "5.54.1",
+    "@typescript-eslint/parser": "5.54.1",
+    "electron": "23.1.3",
     "electron-builder": "23.6.0",
     "electron-devtools-installer": "3.2.0",
-    "eslint": "8.32.0",
-    "eslint-config-prettier": "8.6.0",
+    "eslint": "8.35.0",
+    "eslint-config-prettier": "8.7.0",
     "eslint-plugin-jest": "27.2.1",
-    "eslint-plugin-react": "7.32.0",
+    "eslint-plugin-react": "7.32.2",
     "eslint-plugin-react-hooks": "4.6.0",
-    "eslint-plugin-testing-library": "5.9.1",
-    "globals": "13.19.0",
-    "hardhat": "2.12.3",
-    "husky": "8.0.2",
-    "jest": "29.4.0",
-    "jest-environment-jsdom": "29.4.0",
-    "lint-staged": "13.1.0",
+    "eslint-plugin-testing-library": "5.10.2",
+    "globals": "13.20.0",
+    "hardhat": "2.13.0",
+    "husky": "8.0.3",
+    "jest": "29.5.0",
+    "jest-environment-jsdom": "29.5.0",
+    "lint-staged": "13.1.2",
     "nock": "13.3.0",
-    "node-abi": "3.31.0",
+    "node-abi": "3.33.0",
     "node-gyp": "9.3.1",
     "node-watch": "0.7.3",
     "parcel": "2.8.3",
     "prebuild-install": "7.1.1",
-    "prettier": "2.8.3",
-    "rimraf": "4.1.2",
+    "prettier": "2.8.4",
+    "rimraf": "4.4.0",
     "shx": "0.3.4",
     "stylus": "0.59.0",
     "tree-kill": "1.2.2",
     "ts-jest": "29.0.5",
-    "typescript": "4.9.4",
+    "typescript": "4.9.5",
     "wait-on": "7.0.1"
   },
   "alias": {
@@ -190,8 +191,8 @@
     "url": false
   },
   "engines": {
-    "electron": "22",
-    "node": ">= 16.15.0"
+    "electron": "23",
+    "node": ">= 18.12.1"
   },
   "browserslist": [
     "last 1 Electron major version"
