@@ -1,11 +1,11 @@
-import React, { useState, useEffect } from 'react'
+import { Component } from 'react'
 import Restore from 'react-restore'
 
 import link from '../../../resources/link'
 import Dropdown from '../../../resources/Components/Dropdown'
 import { getSummonShortcut } from '../../../resources/app'
 
-class Settings extends React.Component {
+class Settings extends Component {
   constructor(props, context) {
     super(props, context)
     const latticeEndpoint = context.store('main.latticeSettings.endpointCustom')
@@ -186,7 +186,7 @@ class Settings extends React.Component {
               />
             </div>
             <div className='signerPermissionDetails'>
-              <span>Set Frame's visual theme</span>
+              <span>Set Frame&apos;s visual theme</span>
             </div>
           </div>
 
