@@ -4,7 +4,7 @@ import link from '../../../../../resources/link'
 import svg from '../../../../../resources/svg'
 import { matchFilter } from '../../../../../resources/utils'
 
-import { ClusterBox, Cluster, ClusterRow, ClusterValue } from '../../../../../resources/Components/Cluster'
+import { Cluster, ClusterRow, ClusterValue } from '../../../../../resources/Components/Cluster'
 
 class Inventory extends React.Component {
   constructor(...args) {
