@@ -1,5 +1,4 @@
 import * as Sentry from '@sentry/electron'
-import React from 'react'
 import { createRoot } from 'react-dom/client'
 import Restore from 'react-restore'
 
@@ -16,7 +15,7 @@ document.addEventListener('drop', (e) => e.preventDefault())
 if (process.env.NODE_ENV !== 'development') {
   window.eval = global.eval = () => {
     throw new Error(`This app does not support window.eval()`)
-  } // eslint-disable-line
+  }
 }
 
 function AppComponent() {
