@@ -159,7 +159,7 @@ function identifyKnownContractActions(
 }
 
 const surface = {
-  identity: async (address: string = '', chainId?: number) => {
+  identity: async (address = '', chainId?: number) => {
     // Resolve ens, type and other data about address entities
 
     const results = await Promise.allSettled([
@@ -176,7 +176,7 @@ const surface = {
     return { type, ens }
   },
   resolveEntityType,
-  decode: async (contractAddress: string = '', chainId: number, calldata: string) => {
+  decode: async (contractAddress = '', chainId: number, calldata: string) => {
     // Decode calldata
     const contractSources: ContractSource[] = [{ name: 'ERC-20', source: 'Generic ERC-20', abi: erc20Abi }]
     const contractSource = await fetchContract(contractAddress, chainId)
