@@ -1,7 +1,6 @@
-import React from 'react'
 import useCountdown from '../../Hooks/useCountdown'
 
-export default ({ end, title, titleClass, innerClass }) => {
+const Countdown = ({ end, title, titleClass, innerClass }) => {
   const ttl = isFinite(new Date(end)) ? useCountdown(end) : 'INVALID DATE'
 
   return (
@@ -13,3 +12,5 @@ export default ({ end, title, titleClass, innerClass }) => {
     </div>
   )
 }
+
+export default Countdown
