@@ -1285,7 +1285,7 @@ describe('#navClearSigner', () => {
       }
     ]
 
-    const [req1, req2] = nav
+    const [req1, _req2] = nav
 
     clearSigner('2b')
 
@@ -1330,7 +1330,7 @@ describe('#navClearReq', () => {
       }
     ]
 
-    const [req1, req2, inbox] = nav
+    const [req1, , inbox] = nav
 
     clearRequest('2b')
 
